refactor: cmd options ToOptions method, enforce lll on cmd package (#2776)
* refactor: cmd options ToOptions method, enforce lll on cmd package

* chore: enforce mnd on cmd package

* fix: resolve #2775

* fix: generate parseNodeFlags regression

* chore: enable wsl on cmd package

* chore: appease a few deepsource nits

+++ cmd/context.go
@@ -11,6 +11,10 @@ import (
 	"github.com/spf13/cobra"
 )
 
+const (
+	maxCancelledDestroyTimeout = 20 * time.Second
+)
+
 var onlyOneSignalHandler = make(chan struct{}) //nolint: gochecknoglobals
 
 // SignalHandledContext returns a context that will be canceled if a SIGINT or SIGTERM is
@@ -27,23 +31,35 @@ func SignalHandledContext() (context.Context, context.CancelFunc) {
 
 	go func() {
 		sig := <-sigs
-		log.Errorf("received signal %q, canceling context and cleaning deployment...", sig)
+
+		log.Errorf("received signal %q, canceling context", sig)
 
 		cancel()
 
-		destroyCtx, destroyCancel := context.WithTimeout(context.Background(), 20*time.Second)
+		defer os.Exit(1)
+
+		options := GetOptions()
+
+		if !options.Global.CleanOnCancel {
+			log.Debug("clean on cancel is not true, exiting")
+
+			return
+		}
+
+		destroyCtx, destroyCancel := context.WithTimeout(
+			context.Background(),
+			maxCancelledDestroyTimeout,
+		)
 		defer destroyCancel()
 
 		// destroyFn requires a cobra.Command but only needs the ctx from it
 		destroyCmd := &cobra.Command{}
 		destroyCmd.SetContext(destroyCtx)
 
-		err := destroyFn(destroyCmd, GetOptions())
+		err := destroyFn(destroyCmd, options)
 		if err != nil {
 			log.Errorf("failed destroying lab after cancellation signal: %v", err)
 		}
-
-		os.Exit(1) // skipcq: RVV-A0003
 	}()
 
 	return ctx, cancel
